-список контекстов заменен на синхронизованный -поток чтения консоли перенесен в код сервера -широковещательное сообщения из консоли посылается из кода сервера
