+++ src/ClientCtx.java
@@ -2,24 +2,19 @@ import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.net.Socket;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
+
 
 public class ClientCtx {
-    private Scanner consoleInput;
+
     private DataInputStream inp;
     private DataOutputStream out;
     private Socket socket;
-    private static List<ClientCtx> clients;
 
-    public ClientCtx(Socket socket, List<ClientCtx> clntsList) {
+    public ClientCtx(Socket socket) {
         try {
             this.socket = socket;
-            this.consoleInput = new Scanner(System.in);
             this.inp = new DataInputStream(socket.getInputStream());
             this.out = new DataOutputStream(socket.getOutputStream());
-            this.clients = clntsList;
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -29,7 +24,7 @@ public class ClientCtx {
         return out;
     }
 
-    public void run(ClientCtx clientCtx) {
+    public void run() {
         new Thread(() -> {
             try {
                 while (socket.isConnected()) {
@@ -43,24 +38,7 @@ public class ClientCtx {
                 e.printStackTrace();
             }
         }).start();
-        // отдельный поток для работы с консолью
-        new Thread(() -> {
-                while (socket.isConnected()) {
-                    String msg = consoleInput.nextLine();
-                    System.out.println("Message from Console: " + msg);
-                    notify_all(msg);
-                }
-        }).start();
+
     }
 
-    public static void notify_all(String str) {
-        for (ClientCtx client: clients){
-            try {
-                client.getOut().writeUTF(str);
-                client.getOut().flush();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
 }
